refactors backend structure
