+++ server/src/AppManager.js
@@ -1,25 +1,25 @@
-import { PartyModel, createParty } from "./models/party.js";
+// import { PartyModel, createParty } from "./models/party.js";
 
-export default class AppManager {
-  constructor(socketService, dbService) {
-    this.socketService = socketService;
-  }
+// export default class AppManager {
+//   constructor(socketService, dbService) {
+//     this.socketService = socketService;
+//   }
 
-  async createParty(hostName, settings) {
-    const party = await createParty(settings);
-    this.addPlayerToParty(hostName, party.slug);
-    const socket = this.socketService.create(party.slug);
+//   async createParty(hostName, settings) {
+//     const party = await createParty(settings);
+//     this.addPlayerToParty(hostName, party.slug);
+//     const socket = this.socketService.create(party.slug);
 
-    socket.on("connect", (connection) => {
-      connection.emit("update");
-    });
-    return { slug: party.slug };
-  }
+//     socket.on("connect", (connection) => {
+//       connection.emit("update");
+//     });
+//     return { slug: party.slug };
+//   }
 
-  addPlayerToParty(playerName, partyId) {}
+//   addPlayerToParty(playerName, partyId) {}
 
-  async getParties() {
-    const parties = await PartyModel.find();
-    return parties;
-  }
-}
+//   async getParties() {
+//     const parties = await PartyModel.find();
+//     return parties;
+//   }
+// }
