+++ server/src/db/methods/index.js
@@ -1,0 +1,6 @@
+import devMethods from "./dev";
+import gameMethods from "./game";
+import helpers from "./helpers";
+import partyMethods from "./party";
+
+export { devMethods, gameMethods, helpers, partyMethods };
