+++ server/src/tests/leaveParty.js
@@ -1,6 +1,8 @@
 import mongoose from "mongoose";
 
-import { createParty, clearParties, addPrompt } from "../models/party";
+import { partyMethods, devMethods, gameMethods } from "../db/methods";
+
+const { clearParties } = devMethods;
 
 export const leavePartyTests = () => {
   beforeAll(async () => {
