+++ server/src/tests/renameTeam.js
@@ -1,10 +1,10 @@
 import mongoose from "mongoose";
 
-import {
-  clearParties,
-  createInProgressGame,
-  renameTeam,
-} from "../models/party";
+import { gameMethods, devMethods, helpers } from "../db/methods";
+
+const { clearParties } = devMethods;
+const { renameTeam } = gameMethods;
+const { createInProgressGame } = helpers;
 
 export const renameTeamTests = () => {
   beforeAll(async () => {
@@ -21,16 +21,15 @@ export const renameTeamTests = () => {
 
   it("should rename team at any point during the game", async () => {
     const party = await createInProgressGame("midGame");
-  
+
     const updatedParty = await renameTeam({
       slug: party.slug,
       teamIndex: 0,
       teamName: "Jack's All Star Cat Parade",
     });
 
-    const updatedPartyTeamName = updatedParty.games[0].teams[0].teamName 
+    const updatedPartyTeamName = updatedParty.games[0].teams[0].teamName;
     expect(updatedPartyTeamName).toEqual("Jack's All Star Cat Parade");
-
   });
 
   it("should not rename your team after the game is finished", async () => {
@@ -45,7 +44,4 @@ export const renameTeamTests = () => {
     const updatedPartyTeamName = updatedParty.games[0].teams[0].teamName;
     expect(updatedPartyTeamName).toEqual(party.games[0].teams[0].teamName);
   });
-  
 };
-  
-
