+++ server/src/tests/updateSettings.js
@@ -1,12 +1,10 @@
 import mongoose from "mongoose";
 
-import {
-  createParty,
-  clearParties,
-  addPrompt,
-  createInProgressGame,
-  updateSettings,
-} from "../models/party";
+import { partyMethods, devMethods, helpers } from "../db/methods";
+
+const { clearParties } = devMethods;
+const { updateSettings } = partyMethods;
+const { createInProgressGame } = helpers;
 
 export const updateSettingsTests = () => {
   beforeAll(async () => {
@@ -46,7 +44,7 @@ export const updateSettingsTests = () => {
 
     expect(settings.rotations).toEqual(1000);
   });
-  
+
   it("it should update settings in post game lobby", async () => {
     const party = await createInProgressGame("postGame");
